# apps/core/docs/XWebStack_swagger.yaml
basePath: /apis
definitions:
  ResponseError:
    properties:
      message:
        type: string
    type: object
  dataAccess.DataRefType:
    enum:
    - 0
    - 1
    - 2
    - 4
    - 8
    - 16
    - 32
    type: integer
    x-enum-varnames:
    - TypeUnknown
    - TypeInt
    - TypeFloat
    - TypeDouble
    - TypeFloatArray
    - TypeIntArray
    - TypeData
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.DatarefValue:
    properties:
      dataref_type:
        $ref: '#/definitions/dataAccess.DataRefType'
      name:
        type: string
      value: {}
    type: object
  models.FlightInfo:
    properties:
      airportId:
        type: string
      airportName:
        type: string
      fuelWeight:
        type: number
      time:
        type: number
      totalWeight:
        type: number
    type: object
  models.FlightStatus:
    properties:
      aircraftDisplayName:
        type: string
      aircraftICAO:
        type: string
      arrivalFlightInfo:
        $ref: '#/definitions/models.FlightInfo'
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      departureFlightInfo:
        $ref: '#/definitions/models.FlightInfo'
      events:
        items:
          $ref: '#/definitions/models.FlightStatusEvent'
        type: array
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  models.FlightStatusEvent:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      flightId:
        type: integer
      id:
        type: integer
      timestamp:
        type: number
      updatedAt:
        type: string
    type: object
info:
  contact: {}
paths:
  /flight-logs:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.FlightStatus'
            type: array
        "501":
          description: Not Implemented
      summary: Get a list of FlightLogs
      tags:
      - Flight_Logs
  /flight-logs/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id of a flight log item
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FlightStatus'
        "501":
          description: Not Implemented
      summary: Get one FlightLogs
      tags:
      - Flight_Logs
  /xplm/dataref:
    get:
      consumes:
      - application/json
      parameters:
      - description: xplane dataref string
        in: query
        name: dataref_str
        required: true
        type: string
      - description: alias name, if not set, dataref_str will be used
        in: query
        name: alias
        type: string
      - description: '-1: raw, 2: round up to two digits'
        in: query
        name: precision
        required: true
        type: integer
      - description: transform xplane byte array to string
        in: query
        name: is_byte_array
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DatarefValue'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ResponseError'
      summary: Get Dataref
      tags:
      - XPLM_Dataref
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
      summary: Set Dataref
      tags:
      - XPLM_Dataref
  /xplm/datarefs:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DatarefValue'
            type: array
        "501":
          description: Not Implemented
      summary: Get a list of Dataref
      tags:
      - XPLM_Dataref
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
      summary: Set a list of Dataref
      tags:
      - XPLM_Dataref
swagger: "2.0"